flytt funksjonen til scroll-to.ts
