+++ src/utils/scroll-to.ts
@@ -21,3 +21,15 @@ export const smoothScrollToTarget = (targetId: string, offset = 0) => {
         targetElement.focus({ preventScroll: true });
     }
 };
+
+export const handleStickyScrollOffset = (isOpening: boolean, current: HTMLDivElement | null) => {
+    if (!isOpening && current) {
+        const verticalPosition = current.getBoundingClientRect().top;
+        if (verticalPosition < 0) {
+            window.scrollBy({
+                top: verticalPosition,
+                behavior: 'instant',
+            });
+        }
+    }
+};
